Add tests to pkg/boot/universalpayload
utilities_test.go: tests for functions: `utilities.GetFdtInfo()` (mock DTB file data), `utilities.getPhysicalAddressSizes()` (mock /proc/cpuinfo data)

universalpayload_test.go: tests for constructing Hob list needed by universal payload

+++ pkg/boot/universalpayload/handoffblock.go
@@ -228,7 +228,7 @@ func hobCreateEfiHobCPU() (*EfiHobCPU, error) {
 			HobType:   EfiHobTypeCPU,
 			HobLength: uint16(unsafe.Sizeof(EfiHobCPU{})),
 		},
-		SizeOfMemorySpace: uint8(phyAddrSize),
+		SizeOfMemorySpace: phyAddrSize,
 		SizeOfIOSpace:     DefaultIOAddressSize,
 	}, nil
 }
